CLI tool for bulk decoding
