+++ docker-compose-bind.yml
@@ -1,6 +1,7 @@
 # The compose file for setting up a DNS
 # Web configuration is on https://localhost:10000
 # You must configure a master zone to kingdomhearts.com to your local IP
+# username: root, password: SecretPassword
 
 bind:
   image: sameersbn/bind:latest
