+++ package.json
@@ -4,14 +4,19 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "start": "node index.js --mitm",
+    "decodeJson": "node index.js --decode"
   },
   "author": "",
   "license": "ISC",
   "dependencies": {
     "aes-js": "^3.1.2",
+    "command-line-args": "^5.1.1",
     "http-proxy": "^1.18.1",
     "js-base64": "^2.6.2",
     "ssl-root-cas": "^1.3.1"
+  },
+  "devDependencies": {
+    "eslint": "^7.29.0"
   }
 }
